Moved around some files for an easier workflow
